- batch coarse-stage search + persist steering cache (−30–40% music time)
- Cache coarse/fine steering across frames; “setup” now ~0 ms
- No accuracy change; diagonal loading and logging preserved
- Observed “music” ~0.05–0.07s → ~0.03–0.04s (avg ~0.17–0.18 ms/det)